Don't require both appName and AUMI for WinToastHelper
Let the native library do less magic concerning AUMI and app shortcuts.
Instead give more control to the user of the library and expect them to
do the required checks that a shorcut is available for the desired AUMI
and do not attempt to create shortcuts automatically. Also add more
interactive tests that demonstrate usage and can be used for trying
things.